Cast predicted BW images to float in movie util

+++ problems.py
@@ -194,7 +194,7 @@ class VideoGenSolution:
 
             # threshold prediction in binary case
             if self.cspace == "BW":
-                yhat = yhat > 0.5
+                yhat = (yhat > 0.5).astype('float')
 
             # advance time
             imgs.popleft()
